# cmd/vulnerability-db-api-cli/gen/http/openapi3.yaml
openapi: 3.0.3
info:
  title: Vulnerability Database API
  description: API for the Vulnerability Database
  version: "1.0"
servers:
- url: https://vulnerabilitydb.vulcan.com
paths:
...
  /labels:
    get:
      tags:
      - labels
      summary: List labels labels
      operationId: labels#List labels
      parameters:
      - name: tag
        in: query
        description: Tag of associated targets
        allowEmptyValue: true
        schema:
          type: string
          description: Tag of associated targets
          example: Corrupti facere et architecto deleniti.
        example: Dolores aperiam ut.
      - name: status
        in: query
        description: Status of associated findings
        allowEmptyValue: true
        schema:
          type: string
          description: Status of associated findings
          example: Iusto commodi commodi at sapiente animi dolores.
        example: Molestias reprehenderit ea sunt reprehenderit nemo.
      - name: atDate
        in: query
        description: Allows to get the list of associated findings for a specific
          date (YYY-MM-DD). This filter has preference over minDate and maxDate.
        allowEmptyValue: true
        schema:
          type: string
          description: Allows to get the list of associated findings for a specific
            date (YYY-MM-DD). This filter has preference over minDate and maxDate.
          example: Atque odio ipsam est perspiciatis.
        example: Adipisci rerum earum dolores ut.
      - name: minDate
        in: query
        description: Minimum date filter for associated findings (YYY-MM-DD). This
          filter only applies when used in conjunction with status filter.
        allowEmptyValue: true
        schema:
          type: string
          description: Minimum date filter for associated findings (YYY-MM-DD). This
            filter only applies when used in conjunction with status filter.
          example: Natus consequatur voluptas culpa eum.
        example: Eaque eos nihil.
      - name: maxDate
        in: query
        description: Maximum date filter for associated findings (YYY-MM-DD). This
          filter only applies when used in conjunction with status filter.
        allowEmptyValue: true
        schema:
          type: string
          description: Maximum date filter for associated findings (YYY-MM-DD). This
            filter only applies when used in conjunction with status filter.
          example: Quaerat et placeat quae.
        example: Soluta reprehenderit impedit error odio sunt ut.
      - name: identifiers
        in: query
        description: A list of asset identifiers to filter by for associated targets
        allowEmptyValue: true
        schema:
          type: string
          description: A list of asset identifiers to filter by for associated targets
          example: Minus repellat est ut iure.
        example: Molestiae doloremque.
      responses:
        "200":
          description: OK response.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Labels'
              example:
                labels:
                - Web
                - HTTP
                - TLS
...
  /stats/exposure:
    get:
      tags:
      - stats
      summary: Get stats exposure stats
      operationId: stats#Get stats exposure
      parameters:
      - name: tag
        in: query
        description: Target's tag filter
        allowEmptyValue: true
        schema:
          type: string
          description: Target's tag filter
          example: Quo fugit quos.
        example: Quia illo.
      - name: atDate
        in: query
        description: Allows to get stats at a specific date (YYY-MM-DD)
        allowEmptyValue: true
        schema:
          type: string
          description: Allows to get stats at a specific date (YYY-MM-DD)
          example: Facere quia dolor.
        example: Quas nam dolorem sunt necessitatibus numquam aut.
      - name: minScore
        in: query
        description: Minimum issues score filter
        allowEmptyValue: true
        schema:
          type: integer
          description: Minimum issues score filter
          example: 1396686146028046689
          format: int64
        example: 5632741418978663494
      - name: maxScore
        in: query
        description: Maximum issues score filter
        allowEmptyValue: true
        schema:
          type: integer
          description: Maximum issues score filter
          example: 989680388132904663
          format: int64
        example: 4816450964236605390
      responses:
        "200":
          description: OK response.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Exposure'
              example:
                mean: 0.6576921
                percentile_10: 0.9820335
                percentile_25: 0.5965868
                percentile_50: 0.64732856
                percentile_75: 0.8767213
                percentile_90: 0.9330695
  /stats/fixed:
    get:
      tags:
      - stats
      summary: Get fixed issues stats stats
      description: |-
        This method gets statistics for fixed findings by using different filters.
        When querying for a specific time range, it returns every finding that has been fixed in that time and it's fixed at the end of the time range specified, so a finding that is fixed and then reopened in that range of time will not be counted.
      operationId: stats#Get fixed issues stats
      parameters:
      - name: tag
        in: query
        description: Target's tag filter
        allowEmptyValue: true
        schema:
          type: string
          description: Target's tag filter
          example: Minima doloribus.
        example: Ut porro.
      - name: minDate
        in: query
        description: Minimum date filter (YYY-MM-DD)
        allowEmptyValue: true
        schema:
          type: string
          description: Minimum date filter (YYY-MM-DD)
          example: Placeat accusamus iusto aspernatur rerum dolore quisquam.
        example: Reprehenderit deleniti quia.
      - name: maxDate
        in: query
        description: Maximum date filter (YYY-MM-DD)
        allowEmptyValue: true
        schema:
          type: string
          description: Maximum date filter (YYY-MM-DD)
          example: Qui aut voluptatibus excepturi ut.
        example: Animi et ut qui similique nemo.
      - name: atDate
        in: query
        description: Allows to get stats for fixed findings at a specific date (YYY-MM-DD).
          This filter has preference over minDate and maxDate.
        allowEmptyValue: true
        schema:
          type: string
          description: Allows to get stats for fixed findings at a specific date (YYY-MM-DD).
            This filter has preference over minDate and maxDate.
          example: Asperiores aspernatur commodi.
        example: Qui soluta ut molestiae harum tenetur.
      - name: identifiers
        in: query
        description: A comma separated list of asset identifiers
        allowEmptyValue: true
        schema:
          type: string
          description: A comma separated list of asset identifiers
          example: Aliquam earum dolorum sit ipsa nihil.
        example: Vero voluptates et quos est vero reprehenderit.
      - name: labels
        in: query
        description: A comma separated list of associated labels
        allowEmptyValue: true
        schema:
          type: string
          description: A comma separated list of associated labels
          example: In magnam necessitatibus.
        example: Rerum neque alias.
      responses:
        "200":
          description: OK response.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Severity'
              example:
                critical: 2675288203878151451
                high: 223400334070180172
                informational: 5479921444577369098
                low: 1651328171339960592
                medium: 6156126115486325474
  /stats/issues:
    get:
      tags:
      - stats
      summary: Get issues stats stats
      operationId: stats#Get issues stats
      parameters:
      - name: minDate
        in: query
        description: Minimum date filter (YYY-MM-DD)
        allowEmptyValue: true
        schema:
          type: string
          description: Minimum date filter (YYY-MM-DD)
          example: Praesentium ut ad natus iusto ducimus suscipit.
        example: Quasi eveniet dolorem.
      - name: maxDate
        in: query
        description: Maximum date filter (YYY-MM-DD)
        allowEmptyValue: true
        schema:
          type: string
          description: Maximum date filter (YYY-MM-DD)
          example: Reprehenderit eos et.
        example: Ex eveniet a corrupti beatae ullam.
      - name: minScore
        in: query
        description: Minimum issues score filter
        allowEmptyValue: true
        schema:
          type: integer
          description: Minimum issues score filter
          example: 425485945796897277
          format: int64
        example: 2802762387189910664
      - name: maxScore
        in: query
        description: Maximum issues score filter
        allowEmptyValue: true
        schema:
          type: integer
          description: Maximum issues score filter
          example: 1287179382248988625
          format: int64
        example: 5734992954774312647
      responses:
        "200":
          description: OK response.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/IssueStatsCollection'
              example:
              - cwe_id: 1092913069227277845
                description: |-
                  The remote service accepts connections encrypted using TLS 1.0. TLS 1.0 has a
                  number of cryptographic design flaws. Modern implementations of TLS 1.0
                  mitigate these problems, but newer versions of TLS like 1.1 and 1.2 are
                  designed against these flaws and should be used whenever possible.

                  PCI DSS v3.2 requires that TLS 1.0 be disabled entirely by June 30,
                  2018, except for POS POI terminals (and the SSL/TLS termination
                  points to which they connect) that can be verified as not being
                  susceptible to any known exploits.
                fixed: 3
                id: 234ec73f-74c8-400f-be73-27dae8e5ff16
                open: 1
                recommendations:
                - Enable support for TLS 1.1 and 1.2, and disable support for TLS
                  1.0.
                reference_links:
                - http://www.example.com
                summary: TLS Version 1.0 Protocol Detection
              - cwe_id: 1092913069227277845
                description: |-
                  The remote service accepts connections encrypted using TLS 1.0. TLS 1.0 has a
                  number of cryptographic design flaws. Modern implementations of TLS 1.0
                  mitigate these problems, but newer versions of TLS like 1.1 and 1.2 are
                  designed against these flaws and should be used whenever possible.

                  PCI DSS v3.2 requires that TLS 1.0 be disabled entirely by June 30,
                  2018, except for POS POI terminals (and the SSL/TLS termination
                  points to which they connect) that can be verified as not being
                  susceptible to any known exploits.
                fixed: 3
                id: 234ec73f-74c8-400f-be73-27dae8e5ff16
                open: 1
                recommendations:
                - Enable support for TLS 1.1 and 1.2, and disable support for TLS
                  1.0.
                reference_links:
                - http://www.example.com
                summary: TLS Version 1.0 Protocol Detection
  /stats/mttr:
    get:
      tags:
      - stats
      summary: Get stats MTTR stats
      operationId: stats#Get stats MTTR
      parameters:
      - name: tag
        in: query
        description: Target's tag filter
        allowEmptyValue: true
        schema:
          type: string
          description: Target's tag filter
          example: Voluptatem natus consequuntur minima illo fuga.
        example: Laboriosam voluptatem commodi labore iste necessitatibus ut.
      - name: minDate
        in: query
        description: Minimum date filter (YYY-MM-DD)
        allowEmptyValue: true
        schema:
          type: string
          description: Minimum date filter (YYY-MM-DD)
          example: Sed sint odit molestias cumque.
        example: Ut animi expedita.
      - name: maxDate
        in: query
        description: Maximum date filter (YYY-MM-DD)
        allowEmptyValue: true
        schema:
          type: string
          description: Maximum date filter (YYY-MM-DD)
          example: Corporis omnis delectus.
        example: Iure inventore explicabo dolorum molestias.
      responses:
        "200":
          description: OK response.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Severity'
              example:
                critical: 8619551339247722728
                high: 7830955280873127056
                informational: 2907987811377824896
                low: 3125778721672262496
                medium: 1066780224254076187
  /stats/open:
    get:
      tags:
      - stats
      summary: Get open issues stats stats
      description: |-
        This method gets statistics for open findings by using different filters.
        When querying for a specific time range, it returns every finding that has been found in that time, regardless if that finding has also been fixed in that span of time.
      operationId: stats#Get open issues stats
      parameters:
      - name: tag
        in: query
        description: Target's tag filter
        allowEmptyValue: true
        schema:
          type: string
          description: Target's tag filter
          example: Repudiandae est nihil quis.
        example: Ad et ex accusantium eaque.
      - name: minDate
        in: query
        description: Minimum date filter (YYY-MM-DD)
        allowEmptyValue: true
        schema:
          type: string
          description: Minimum date filter (YYY-MM-DD)
          example: Eveniet quia.
        example: Magni aut dolorum dolores.
      - name: maxDate
        in: query
        description: Maximum date filter (YYY-MM-DD)
        allowEmptyValue: true
        schema:
          type: string
          description: Maximum date filter (YYY-MM-DD)
          example: Non consequatur adipisci eum.
        example: Aliquid possimus non aut.
      - name: atDate
        in: query
        description: Allows to get stats for open findings at a specific date (YYY-MM-DD).
          This filter has preference over minDate and maxDate.
        allowEmptyValue: true
        schema:
          type: string
          description: Allows to get stats for open findings at a specific date (YYY-MM-DD).
            This filter has preference over minDate and maxDate.
          example: Omnis inventore error et est autem.
        example: Et officiis.
      - name: identifiers
        in: query
        description: A comma separated list of asset identifiers
        allowEmptyValue: true
        schema:
          type: string
          description: A comma separated list of asset identifiers
          example: Enim praesentium corporis.
        example: Officia deserunt fuga.
      - name: labels
        in: query
        description: A comma separated list of associated labels
        allowEmptyValue: true
        schema:
          type: string
          description: A comma separated list of associated labels
          example: Officiis nobis.
        example: Et earum ad eos eius asperiores.
      responses:
        "200":
          description: OK response.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Severity'
              example:
                critical: 2615216803934210996
                high: 2514681863203349240
                informational: 3941938844913984946
                low: 1064693259099015916
                medium: 5392709263805215660
...
components:
  schemas:
...
    Exposure:
      type: object
      properties:
        mean:
          type: number
          description: Mean exposure
          example: 0.0017000929
          format: float
        percentile_10:
          type: number
          description: Percentile 10 of exposure
          example: 0.042889617
          format: float
        percentile_25:
          type: number
          description: Percentile 25 of exposure or first quartile
          example: 0.31601247
          format: float
        percentile_50:
          type: number
          description: Percentile 50 of exposure or median
          example: 0.97426087
          format: float
        percentile_75:
          type: number
          description: Percentile 75 of exposure or third quartile
          example: 0.5328149
          format: float
        percentile_90:
          type: number
          description: Percentile 90 of exposure
          example: 0.847257
          format: float
      example:
        mean: 0.52007186
        percentile_10: 0.7937411
        percentile_25: 0.9130958
        percentile_50: 0.021809133
        percentile_75: 0.39894477
        percentile_90: 0.10781263
...
    IssueStatsCollection:
      type: array
      items:
        $ref: '#/components/schemas/IssueStats'
      example:
      - cwe_id: 1092913069227277845
        description: |-
          The remote service accepts connections encrypted using TLS 1.0. TLS 1.0 has a
          number of cryptographic design flaws. Modern implementations of TLS 1.0
          mitigate these problems, but newer versions of TLS like 1.1 and 1.2 are
          designed against these flaws and should be used whenever possible.

          PCI DSS v3.2 requires that TLS 1.0 be disabled entirely by June 30,
          2018, except for POS POI terminals (and the SSL/TLS termination
          points to which they connect) that can be verified as not being
          susceptible to any known exploits.
        fixed: 3
        id: 234ec73f-74c8-400f-be73-27dae8e5ff16
        open: 1
        recommendations:
        - Enable support for TLS 1.1 and 1.2, and disable support for TLS 1.0.
        reference_links:
        - http://www.example.com
        summary: TLS Version 1.0 Protocol Detection
      - cwe_id: 1092913069227277845
        description: |-
          The remote service accepts connections encrypted using TLS 1.0. TLS 1.0 has a
          number of cryptographic design flaws. Modern implementations of TLS 1.0
          mitigate these problems, but newer versions of TLS like 1.1 and 1.2 are
          designed against these flaws and should be used whenever possible.

          PCI DSS v3.2 requires that TLS 1.0 be disabled entirely by June 30,
          2018, except for POS POI terminals (and the SSL/TLS termination
          points to which they connect) that can be verified as not being
          susceptible to any known exploits.
        fixed: 3
        id: 234ec73f-74c8-400f-be73-27dae8e5ff16
        open: 1
        recommendations:
        - Enable support for TLS 1.1 and 1.2, and disable support for TLS 1.0.
        reference_links:
        - http://www.example.com
        summary: TLS Version 1.0 Protocol Detection
      - cwe_id: 1092913069227277845
        description: |-
          The remote service accepts connections encrypted using TLS 1.0. TLS 1.0 has a
          number of cryptographic design flaws. Modern implementations of TLS 1.0
          mitigate these problems, but newer versions of TLS like 1.1 and 1.2 are
          designed against these flaws and should be used whenever possible.

          PCI DSS v3.2 requires that TLS 1.0 be disabled entirely by June 30,
          2018, except for POS POI terminals (and the SSL/TLS termination
          points to which they connect) that can be verified as not being
          susceptible to any known exploits.
        fixed: 3
        id: 234ec73f-74c8-400f-be73-27dae8e5ff16
        open: 1
        recommendations:
        - Enable support for TLS 1.1 and 1.2, and disable support for TLS 1.0.
        reference_links:
        - http://www.example.com
        summary: TLS Version 1.0 Protocol Detection
      - cwe_id: 1092913069227277845
        description: |-
          The remote service accepts connections encrypted using TLS 1.0. TLS 1.0 has a
          number of cryptographic design flaws. Modern implementations of TLS 1.0
          mitigate these problems, but newer versions of TLS like 1.1 and 1.2 are
          designed against these flaws and should be used whenever possible.

          PCI DSS v3.2 requires that TLS 1.0 be disabled entirely by June 30,
          2018, except for POS POI terminals (and the SSL/TLS termination
          points to which they connect) that can be verified as not being
          susceptible to any known exploits.
        fixed: 3
        id: 234ec73f-74c8-400f-be73-27dae8e5ff16
        open: 1
        recommendations:
        - Enable support for TLS 1.1 and 1.2, and disable support for TLS 1.0.
        reference_links:
        - http://www.example.com
        summary: TLS Version 1.0 Protocol Detection
...
    Labels:
      type: object
      properties:
        labels:
          type: array
          items:
            type: string
            example: Dignissimos molestias unde rerum.
          description: List of labels
          example:
          - Web
          - HTTP
          - TLS
      example:
        labels:
        - Web
        - HTTP
        - TLS
...
    Severity:
      type: object
      properties:
        critical:
          type: integer
          description: Number of critial issues
          example: 7740579150252222078
          format: int64
        high:
          type: integer
          description: Number of high issues
          example: 1924022982642599779
          format: int64
        informational:
          type: integer
          description: Number of informational issues
          example: 7024809593536588029
          format: int64
        low:
          type: integer
          description: Number of low issues
          example: 2755035232482032889
          format: int64
        medium:
          type: integer
          description: Number of medium issues
          example: 388146050375883564
          format: int64
      example:
        critical: 2455764659827022788
        high: 1365278855829983458
        informational: 6695950340618226718
        low: 675546947352136462
        medium: 206070019598881958
...
  securitySchemes:
    authorization_header_Authorization:
      type: apiKey
      description: Value must be 'TAG tag=mytag'
      name: Authorization
      in: header
...